加入password raw

+++ sched/src/main/java/com/nb6868/onex/uc/entity/TaskLogEntity.java
@@ -48,7 +48,7 @@ public class TaskLogEntity extends BaseEntity {
     /**
      * 耗时(单位：毫秒)
      */
-	private Integer times;
+	private Long times;
     /**
      * 租户编码
      */
